+++ src/main/java/gov/nist/csd/pm/core/impl/grpc/client/GrpcAccessQuerier.java
@@ -162,6 +162,6 @@ public class GrpcAccessQuerier implements AccessQuery {
 
     @Override
     public Map<Long, Set<Long>> computeRequiredAttributeSets(TargetContext targetCtx, AccessRightSet privileges) {
-        return null;//TODO
+
     }
 }
