+++ src/main/java/gov/nist/csd/pm/core/pap/graph/dag/NoopShortCircuit.java
@@ -1,4 +1,4 @@
-package gov.nist.csd.pm.core.common.graph.dag;
+package gov.nist.csd.pm.core.pap.graph.dag;
 
 public class NoopShortCircuit implements ShortCircuit{
     @Override
