+++ src/main/java/gov/nist/csd/pm/core/pap/query/model/context/TargetContext.java
@@ -1,7 +1,7 @@
 package gov.nist.csd.pm.core.pap.query.model.context;
 
 import gov.nist.csd.pm.core.common.exception.PMException;
-import gov.nist.csd.pm.core.common.graph.dag.GraphWalker;
+import gov.nist.csd.pm.core.pap.graph.dag.GraphWalker;
 import java.io.Serializable;
 import java.util.Collection;
 import java.util.Set;
