+++ src/test/java/gov/nist/csd/pm/core/impl/memory/pap/dag/DepthFirstGraphWalkerTest.java
@@ -3,7 +3,7 @@ package gov.nist.csd.pm.core.impl.memory.pap.dag;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import gov.nist.csd.pm.core.common.exception.PMException;
-import gov.nist.csd.pm.core.common.graph.dag.GraphWalker;
+import gov.nist.csd.pm.core.pap.graph.dag.GraphWalker;
 import gov.nist.csd.pm.core.pap.PAP;
 import gov.nist.csd.pm.core.pap.graph.dag.DepthFirstGraphWalker;
 import gov.nist.csd.pm.core.util.TestPAP;
